+++ src/lib/constants.js
@@ -1,8 +1,8 @@
 module.exports = {
   DECK_LIST_TYPES: [
     'mainDeck',
-    'sideboard',
-    'commandZone'
+    'commandZone',
+    'sideboard'
   ],
   MISSING_CARD_IMAGE: 'https://img.scryfall.com/errors/missing.jpg'
 }